refactor: remove auto-deactivation, add dedicated service provider and update notice
